work on basic code structure
