+++ src/NRS.jl
@@ -1,2 +1,11 @@
-module NRS 
+module NRS
+import Random 
+import SparseArrayKit
+import TensorOperations
+
+include("types.jl")
+include("net.jl")
+include("rewrite.jl")
+include("symbols.jl")
+include("utils.jl")
 end
