+++ src/net.jl
@@ -1,15 +1,28 @@
-function detect_conflicts(net::Net) end
 
-function resolve_conflicts!(net::Net, nrs::NRS, conflicts::Vector{Tuple{Int,Int}}; rng::Random.AbstractRNG=Random.GLOBAL_RNG) end
+struct Net{T<:AbstractArray,M<:AbstractArray} <: AbstractNet
+    inputs::T
+    outputs::T
+    markings::M
+    enabled::Vector{Bool}
+end
 
-function compute_enabled!(net::Net) 
+function detect_conflicts(net::Net)
+    # TODO:
+end
+
+function resolve_conflicts!(net::Net, conflicts::Vector{Tuple{Int,Int}};
+                            rng::Random.AbstractRNG=Random.GLOBAL_RNG)
+    # TODO:
+end
+
+function compute_enabled!(net::Net)
     # TODO: add tensor math for conflict
 end
 
-function step!(net::Net; rng::Random.AbstractRNG=Random.GLOBAL_RNG) 
+function step!(net::Net; rng::Random.AbstractRNG=Random.GLOBAL_RNG)
     compute_enabled!(net)
     conflicts = detect_conflicts(net)
-    resolve_conflicts!(net, conflicts, rng)
+    return resolve_conflicts!(net, conflicts; rng=rng)
 
-    # TODO add tensor math for step
+    # TODO: add tensor math for step
 end
